move config.rs files to src/config
