+++ src/config/rocksdb.rs
@@ -1,47 +1,42 @@
 // Copyright 2017 TiKV Project Authors. Licensed under Apache-2.0.
 
-//! Configuration for the entire server.
-//!
-//! TiKV is configured through the `TiKvConfig` type, which is in turn
-//! made up of many other configuration types.
-
-use std::cmp::{self, Ord, Ordering};
+// use std::cmp::{self, Ord, Ordering};
 use std::error::Error;
-use std::fs;
-use std::i32;
-use std::io::Error as IoError;
+// use std::fs;
+// use std::i32;
+// use std::io::Error as IoError;
 use std::io::Write;
-use std::path::Path;
-use std::sync::Arc;
+// use std::path::Path;
+// use std::sync::Arc;
 use std::time::Duration;
 use std::usize;
 
-use kvproto::configpb::{self, StatusCode};
+// use kvproto::configpb::{self, StatusCode};
 
-use configuration::{ConfigChange, ConfigValue, Configuration};
+// use configuration::{ConfigChange, ConfigValue, Configuration};
 use engine::rocks::{
     BlockBasedOptions, Cache, ColumnFamilyOptions, CompactionPriority, DBCompactionStyle,
     DBCompressionType, DBOptions, DBRateLimiterMode, DBRecoveryMode, LRUCacheOptions,
     TitanDBOptions,
 };
-use slog;
+// use slog;
 use sys_info;
 
-use crate::import::Config as ImportConfig;
+// use crate::import::Config as ImportConfig;
 use crate::raftstore::coprocessor::properties::{
     MvccPropertiesCollectorFactory, RangePropertiesCollectorFactory,
 };
 use crate::raftstore::coprocessor::properties::{
     DEFAULT_PROP_KEYS_INDEX_DISTANCE, DEFAULT_PROP_SIZE_INDEX_DISTANCE,
 };
-use crate::raftstore::coprocessor::Config as CopConfig;
-use crate::raftstore::store::Config as RaftstoreConfig;
-use crate::raftstore::store::PdTask;
-use crate::server::gc_worker::GcConfig;
-use crate::server::lock_manager::Config as PessimisticTxnConfig;
-use crate::server::Config as ServerConfig;
+// use crate::raftstore::coprocessor::Config as CopConfig;
+// use crate::raftstore::store::Config as RaftstoreConfig;
+// use crate::raftstore::store::PdTask;
+// use crate::server::gc_worker::GcConfig;
+// use crate::server::lock_manager::Config as PessimisticTxnConfig;
+// use crate::server::Config as ServerConfig;
 use crate::server::CONFIG_ROCKSDB_GAUGE;
-use crate::storage::config::{Config as StorageConfig, DEFAULT_DATA_DIR, DEFAULT_ROCKSDB_SUB_DIR};
+// use crate::storage::config::{Config as StorageConfig, DEFAULT_DATA_DIR, DEFAULT_ROCKSDB_SUB_DIR};
 use engine::rocks::util::config::{self as rocks_config, BlobRunMode, CompressionType};
 use engine::rocks::util::{
     db_exist, CFOptions, EventListener, FixedPrefixSliceTransform, FixedSuffixSliceTransform,
@@ -49,12 +44,7 @@ use engine::rocks::util::{
 };
 use engine::{CF_DEFAULT, CF_LOCK, CF_RAFT, CF_WRITE};
 use keys::region_raft_prefix_len;
-use pd_client::{Config as PdConfig, PdClient};
 use tikv_util::config::{self, ReadableDuration, ReadableSize, GB, KB, MB};
-use tikv_util::future_pool;
-use tikv_util::security::SecurityConfig;
-use tikv_util::time::duration_to_sec;
-use tikv_util::worker::FutureScheduler;
 
 const LOCKCF_MIN_MEM: usize = 256 * MB as usize;
 const LOCKCF_MAX_MEM: usize = GB as usize;
@@ -1054,1034 +1044,5 @@ impl RaftDbConfig {
             }
         }
         Ok(())
-    }
-}
-
-#[derive(Clone, Serialize, Deserialize, PartialEq, Debug)]
-#[serde(default)]
-#[serde(rename_all = "kebab-case")]
-pub struct MetricConfig {
-    pub interval: ReadableDuration,
-    pub address: String,
-    pub job: String,
-}
-
-impl Default for MetricConfig {
-    fn default() -> MetricConfig {
-        MetricConfig {
-            interval: ReadableDuration::secs(15),
-            address: "".to_owned(),
-            job: "tikv".to_owned(),
-        }
-    }
-}
-
-pub mod log_level_serde {
-    use serde::{
-        de::{Error, Unexpected},
-        Deserialize, Deserializer, Serialize, Serializer,
-    };
-    use slog::Level;
-    use tikv_util::logger::{get_level_by_string, get_string_by_level};
-
-    pub fn deserialize<'de, D>(deserializer: D) -> Result<Level, D::Error>
-    where
-        D: Deserializer<'de>,
-    {
-        let string = String::deserialize(deserializer)?;
-        get_level_by_string(&string)
-            .ok_or_else(|| D::Error::invalid_value(Unexpected::Str(&string), &"a valid log level"))
-    }
-
-    #[allow(clippy::trivially_copy_pass_by_ref)]
-    pub fn serialize<S>(value: &Level, serializer: S) -> Result<S::Ok, S::Error>
-    where
-        S: Serializer,
-    {
-        get_string_by_level(*value).serialize(serializer)
-    }
-}
-
-macro_rules! readpool_config {
-    ($struct_name:ident, $test_mod_name:ident, $display_name:expr) => {
-        #[derive(Clone, Copy, Serialize, Deserialize, PartialEq, Debug)]
-        #[serde(default)]
-        #[serde(rename_all = "kebab-case")]
-        pub struct $struct_name {
-            pub high_concurrency: usize,
-            pub normal_concurrency: usize,
-            pub low_concurrency: usize,
-            pub max_tasks_per_worker_high: usize,
-            pub max_tasks_per_worker_normal: usize,
-            pub max_tasks_per_worker_low: usize,
-            pub stack_size: ReadableSize,
-        }
-
-        impl $struct_name {
-            /// Builds configurations for low, normal and high priority pools.
-            pub fn to_future_pool_configs(&self) -> Vec<future_pool::Config> {
-                vec![
-                    future_pool::Config {
-                        workers: self.low_concurrency,
-                        max_tasks_per_worker: self.max_tasks_per_worker_low,
-                        stack_size: self.stack_size.0 as usize,
-                    },
-                    future_pool::Config {
-                        workers: self.normal_concurrency,
-                        max_tasks_per_worker: self.max_tasks_per_worker_normal,
-                        stack_size: self.stack_size.0 as usize,
-                    },
-                    future_pool::Config {
-                        workers: self.high_concurrency,
-                        max_tasks_per_worker: self.max_tasks_per_worker_high,
-                        stack_size: self.stack_size.0 as usize,
-                    },
-                ]
-            }
-
-            pub fn default_for_test() -> Self {
-                Self {
-                    high_concurrency: 2,
-                    normal_concurrency: 2,
-                    low_concurrency: 2,
-                    max_tasks_per_worker_high: 2000,
-                    max_tasks_per_worker_normal: 2000,
-                    max_tasks_per_worker_low: 2000,
-                    stack_size: ReadableSize::mb(1),
-                }
-            }
-
-            pub fn validate(&self) -> Result<(), Box<dyn Error>> {
-                if self.high_concurrency == 0 {
-                    return Err(format!(
-                        "readpool.{}.high-concurrency should be > 0",
-                        $display_name
-                    )
-                    .into());
-                }
-                if self.normal_concurrency == 0 {
-                    return Err(format!(
-                        "readpool.{}.normal-concurrency should be > 0",
-                        $display_name
-                    )
-                    .into());
-                }
-                if self.low_concurrency == 0 {
-                    return Err(format!(
-                        "readpool.{}.low-concurrency should be > 0",
-                        $display_name
-                    )
-                    .into());
-                }
-                if self.stack_size.0 < ReadableSize::mb(2).0 {
-                    return Err(
-                        format!("readpool.{}.stack-size should be >= 2mb", $display_name).into(),
-                    );
-                }
-                if self.max_tasks_per_worker_high <= 1 {
-                    return Err(format!(
-                        "readpool.{}.max-tasks-per-worker-high should be > 1",
-                        $display_name
-                    )
-                    .into());
-                }
-                if self.max_tasks_per_worker_normal <= 1 {
-                    return Err(format!(
-                        "readpool.{}.max-tasks-per-worker-normal should be > 1",
-                        $display_name
-                    )
-                    .into());
-                }
-                if self.max_tasks_per_worker_low <= 1 {
-                    return Err(format!(
-                        "readpool.{}.max-tasks-per-worker-low should be > 1",
-                        $display_name
-                    )
-                    .into());
-                }
-
-                Ok(())
-            }
-        }
-
-        #[cfg(test)]
-        mod $test_mod_name {
-            use super::*;
-
-            #[test]
-            fn test_validate() {
-                let cfg = $struct_name::default();
-                assert!(cfg.validate().is_ok());
-
-                let mut invalid_cfg = cfg.clone();
-                invalid_cfg.high_concurrency = 0;
-                assert!(invalid_cfg.validate().is_err());
-
-                let mut invalid_cfg = cfg.clone();
-                invalid_cfg.normal_concurrency = 0;
-                assert!(invalid_cfg.validate().is_err());
-
-                let mut invalid_cfg = cfg.clone();
-                invalid_cfg.low_concurrency = 0;
-                assert!(invalid_cfg.validate().is_err());
-
-                let mut invalid_cfg = cfg.clone();
-                invalid_cfg.stack_size = ReadableSize::mb(1);
-                assert!(invalid_cfg.validate().is_err());
-
-                let mut invalid_cfg = cfg.clone();
-                invalid_cfg.max_tasks_per_worker_high = 0;
-                assert!(invalid_cfg.validate().is_err());
-                invalid_cfg.max_tasks_per_worker_high = 1;
-                assert!(invalid_cfg.validate().is_err());
-                invalid_cfg.max_tasks_per_worker_high = 100;
-                assert!(cfg.validate().is_ok());
-
-                let mut invalid_cfg = cfg.clone();
-                invalid_cfg.max_tasks_per_worker_normal = 0;
-                assert!(invalid_cfg.validate().is_err());
-                invalid_cfg.max_tasks_per_worker_normal = 1;
-                assert!(invalid_cfg.validate().is_err());
-                invalid_cfg.max_tasks_per_worker_normal = 100;
-                assert!(cfg.validate().is_ok());
-
-                let mut invalid_cfg = cfg.clone();
-                invalid_cfg.max_tasks_per_worker_low = 0;
-                assert!(invalid_cfg.validate().is_err());
-                invalid_cfg.max_tasks_per_worker_low = 1;
-                assert!(invalid_cfg.validate().is_err());
-                invalid_cfg.max_tasks_per_worker_low = 100;
-                assert!(cfg.validate().is_ok());
-            }
-        }
-    };
-}
-
-const DEFAULT_STORAGE_READPOOL_MIN_CONCURRENCY: usize = 4;
-const DEFAULT_STORAGE_READPOOL_MAX_CONCURRENCY: usize = 8;
-
-// Assume a request can be finished in 1ms, a request at position x will wait about
-// 0.001 * x secs to be actual started. A server-is-busy error will trigger 2 seconds
-// backoff. So when it needs to wait for more than 2 seconds, return error won't causse
-// larger latency.
-const DEFAULT_READPOOL_MAX_TASKS_PER_WORKER: usize = 2 as usize * 1000;
-
-const DEFAULT_READPOOL_STACK_SIZE_MB: u64 = 10;
-
-readpool_config!(StorageReadPoolConfig, storage_read_pool_test, "storage");
-
-impl Default for StorageReadPoolConfig {
-    fn default() -> Self {
-        let cpu_num = sys_info::cpu_num().unwrap();
-        let mut concurrency = (f64::from(cpu_num) * 0.5) as usize;
-        concurrency = cmp::max(DEFAULT_STORAGE_READPOOL_MIN_CONCURRENCY, concurrency);
-        concurrency = cmp::min(DEFAULT_STORAGE_READPOOL_MAX_CONCURRENCY, concurrency);
-        Self {
-            high_concurrency: concurrency,
-            normal_concurrency: concurrency,
-            low_concurrency: concurrency,
-            max_tasks_per_worker_high: DEFAULT_READPOOL_MAX_TASKS_PER_WORKER,
-            max_tasks_per_worker_normal: DEFAULT_READPOOL_MAX_TASKS_PER_WORKER,
-            max_tasks_per_worker_low: DEFAULT_READPOOL_MAX_TASKS_PER_WORKER,
-            stack_size: ReadableSize::mb(DEFAULT_READPOOL_STACK_SIZE_MB),
-        }
-    }
-}
-
-const DEFAULT_COPROCESSOR_READPOOL_MIN_CONCURRENCY: usize = 2;
-
-readpool_config!(
-    CoprReadPoolConfig,
-    coprocessor_read_pool_test,
-    "coprocessor"
-);
-
-impl Default for CoprReadPoolConfig {
-    fn default() -> Self {
-        let cpu_num = sys_info::cpu_num().unwrap();
-        let mut concurrency = (f64::from(cpu_num) * 0.8) as usize;
-        concurrency = cmp::max(DEFAULT_COPROCESSOR_READPOOL_MIN_CONCURRENCY, concurrency);
-        Self {
-            high_concurrency: concurrency,
-            normal_concurrency: concurrency,
-            low_concurrency: concurrency,
-            max_tasks_per_worker_high: DEFAULT_READPOOL_MAX_TASKS_PER_WORKER,
-            max_tasks_per_worker_normal: DEFAULT_READPOOL_MAX_TASKS_PER_WORKER,
-            max_tasks_per_worker_low: DEFAULT_READPOOL_MAX_TASKS_PER_WORKER,
-            stack_size: ReadableSize::mb(DEFAULT_READPOOL_STACK_SIZE_MB),
-        }
-    }
-}
-
-#[derive(Clone, Serialize, Deserialize, PartialEq, Debug, Default)]
-#[serde(default)]
-#[serde(rename_all = "kebab-case")]
-pub struct ReadPoolConfig {
-    pub storage: StorageReadPoolConfig,
-    pub coprocessor: CoprReadPoolConfig,
-}
-
-impl ReadPoolConfig {
-    pub fn validate(&mut self) -> Result<(), Box<dyn Error>> {
-        self.storage.validate()?;
-        self.coprocessor.validate()?;
-        Ok(())
-    }
-}
-
-#[derive(Clone, Serialize, Deserialize, PartialEq, Debug, Configuration)]
-#[serde(default)]
-#[serde(rename_all = "kebab-case")]
-pub struct TiKvConfig {
-    #[config(skip)]
-    #[serde(with = "log_level_serde")]
-    pub log_level: slog::Level,
-    #[config(skip)]
-    pub log_file: String,
-    #[config(skip)]
-    pub log_rotation_timespan: ReadableDuration,
-    #[config(skip)]
-    pub panic_when_unexpected_key_or_data: bool,
-    pub refresh_config_interval: ReadableDuration,
-    #[config(skip)]
-    pub readpool: ReadPoolConfig,
-    #[config(skip)]
-    pub server: ServerConfig,
-    #[config(skip)]
-    pub storage: StorageConfig,
-    #[config(skip)]
-    pub pd: PdConfig,
-    #[config(skip)]
-    pub metric: MetricConfig,
-    #[config(submodule)]
-    #[serde(rename = "raftstore")]
-    pub raft_store: RaftstoreConfig,
-    #[config(skip)]
-    pub coprocessor: CopConfig,
-    #[config(skip)]
-    pub rocksdb: DbConfig,
-    #[config(skip)]
-    pub raftdb: RaftDbConfig,
-    #[config(skip)]
-    pub security: SecurityConfig,
-    #[config(skip)]
-    pub import: ImportConfig,
-    #[config(skip)]
-    pub pessimistic_txn: PessimisticTxnConfig,
-    #[config(skip)]
-    pub gc: GcConfig,
-}
-
-impl Default for TiKvConfig {
-    fn default() -> TiKvConfig {
-        TiKvConfig {
-            log_level: slog::Level::Info,
-            log_file: "".to_owned(),
-            log_rotation_timespan: ReadableDuration::hours(24),
-            panic_when_unexpected_key_or_data: false,
-            refresh_config_interval: ReadableDuration::secs(30),
-            readpool: ReadPoolConfig::default(),
-            server: ServerConfig::default(),
-            metric: MetricConfig::default(),
-            raft_store: RaftstoreConfig::default(),
-            coprocessor: CopConfig::default(),
-            pd: PdConfig::default(),
-            rocksdb: DbConfig::default(),
-            raftdb: RaftDbConfig::default(),
-            storage: StorageConfig::default(),
-            security: SecurityConfig::default(),
-            import: ImportConfig::default(),
-            pessimistic_txn: PessimisticTxnConfig::default(),
-            gc: GcConfig::default(),
-        }
-    }
-}
-
-impl TiKvConfig {
-    pub fn validate(&mut self) -> Result<(), Box<dyn Error>> {
-        self.readpool.validate()?;
-        self.storage.validate()?;
-
-        self.raft_store.region_split_check_diff = self.coprocessor.region_split_size / 16;
-        self.raft_store.raftdb_path = if self.raft_store.raftdb_path.is_empty() {
-            config::canonicalize_sub_path(&self.storage.data_dir, "raft")?
-        } else {
-            config::canonicalize_path(&self.raft_store.raftdb_path)?
-        };
-
-        let kv_db_path =
-            config::canonicalize_sub_path(&self.storage.data_dir, DEFAULT_ROCKSDB_SUB_DIR)?;
-
-        if kv_db_path == self.raft_store.raftdb_path {
-            return Err("raft_store.raftdb_path can not same with storage.data_dir/db".into());
-        }
-        if db_exist(&kv_db_path) && !db_exist(&self.raft_store.raftdb_path) {
-            return Err("default rocksdb exist, buf raftdb not exist".into());
-        }
-        if !db_exist(&kv_db_path) && db_exist(&self.raft_store.raftdb_path) {
-            return Err("default rocksdb not exist, buf raftdb exist".into());
-        }
-
-        // Check blob file dir is empty when titan is disabled
-        if !self.rocksdb.titan.enabled {
-            let titandb_path = if self.rocksdb.titan.dirname.is_empty() {
-                Path::new(&kv_db_path).join("titandb")
-            } else {
-                Path::new(&self.rocksdb.titan.dirname).to_path_buf()
-            };
-            if let Err(e) =
-                tikv_util::config::check_data_dir_empty(titandb_path.to_str().unwrap(), "blob")
-            {
-                return Err(format!(
-                    "check: titandb-data-dir-empty; err: \"{}\"; \
-                     hint: You have disabled titan when its data directory is not empty. \
-                     To properly shutdown titan, please enter fallback blob-run-mode and \
-                     wait till titandb files are all safely ingested.",
-                    e
-                )
-                .into());
-            }
-        }
-
-        let expect_keepalive = self.raft_store.raft_heartbeat_interval() * 2;
-        if expect_keepalive > self.server.grpc_keepalive_time.0 {
-            return Err(format!(
-                "grpc_keepalive_time is too small, it should not less than the double of \
-                 raft tick interval (>= {})",
-                duration_to_sec(expect_keepalive)
-            )
-            .into());
-        }
-
-        self.rocksdb.validate()?;
-        self.raftdb.validate()?;
-        self.server.validate()?;
-        self.raft_store.validate()?;
-        self.pd.validate()?;
-        self.coprocessor.validate()?;
-        self.security.validate()?;
-        self.import.validate()?;
-        self.pessimistic_txn.validate()?;
-        self.gc.validate()?;
-        Ok(())
-    }
-
-    pub fn compatible_adjust(&mut self) {
-        let default_raft_store = RaftstoreConfig::default();
-        let default_coprocessor = CopConfig::default();
-        if self.raft_store.region_max_size != default_raft_store.region_max_size {
-            warn!(
-                "deprecated configuration, \
-                 raftstore.region-max-size has been moved to coprocessor"
-            );
-            if self.coprocessor.region_max_size == default_coprocessor.region_max_size {
-                warn!(
-                    "override coprocessor.region-max-size with raftstore.region-max-size, {:?}",
-                    self.raft_store.region_max_size
-                );
-                self.coprocessor.region_max_size = self.raft_store.region_max_size;
-            }
-            self.raft_store.region_max_size = default_raft_store.region_max_size;
-        }
-        if self.raft_store.region_split_size != default_raft_store.region_split_size {
-            warn!(
-                "deprecated configuration, \
-                 raftstore.region-split-size has been moved to coprocessor",
-            );
-            if self.coprocessor.region_split_size == default_coprocessor.region_split_size {
-                warn!(
-                    "override coprocessor.region-split-size with raftstore.region-split-size, {:?}",
-                    self.raft_store.region_split_size
-                );
-                self.coprocessor.region_split_size = self.raft_store.region_split_size;
-            }
-            self.raft_store.region_split_size = default_raft_store.region_split_size;
-        }
-        if self.server.end_point_concurrency.is_some() {
-            warn!(
-                "deprecated configuration, {} has been moved to {}",
-                "server.end-point-concurrency", "readpool.coprocessor.xxx-concurrency",
-            );
-            warn!(
-                "override {} with {}, {:?}",
-                "readpool.coprocessor.xxx-concurrency",
-                "server.end-point-concurrency",
-                self.server.end_point_concurrency
-            );
-            let concurrency = self.server.end_point_concurrency.unwrap();
-            self.readpool.coprocessor.high_concurrency = concurrency;
-            self.readpool.coprocessor.normal_concurrency = concurrency;
-            self.readpool.coprocessor.low_concurrency = concurrency;
-        }
-        if self.server.end_point_stack_size.is_some() {
-            warn!(
-                "deprecated configuration, {} has been moved to {}",
-                "server.end-point-stack-size", "readpool.coprocessor.stack-size",
-            );
-            warn!(
-                "override {} with {}, {:?}",
-                "readpool.coprocessor.stack-size",
-                "server.end-point-stack-size",
-                self.server.end_point_stack_size
-            );
-            self.readpool.coprocessor.stack_size = self.server.end_point_stack_size.unwrap();
-        }
-        if self.server.end_point_max_tasks.is_some() {
-            warn!(
-                "deprecated configuration, {} is no longer used and ignored, please use {}.",
-                "server.end-point-max-tasks", "readpool.coprocessor.max-tasks-per-worker-xxx",
-            );
-            // Note:
-            // Our `end_point_max_tasks` is mostly mistakenly configured, so we don't override
-            // new configuration using old values.
-            self.server.end_point_max_tasks = None;
-        }
-        if self.raft_store.clean_stale_peer_delay.as_secs() > 0 {
-            let delay_secs = self.raft_store.clean_stale_peer_delay.as_secs()
-                + self.server.end_point_request_max_handle_duration.as_secs();
-            self.raft_store.clean_stale_peer_delay = ReadableDuration::secs(delay_secs);
-        }
-        // When shared block cache is enabled, if its capacity is set, it overrides individual
-        // block cache sizes. Otherwise use the sum of block cache size of all column families
-        // as the shared cache size.
-        let cache_cfg = &mut self.storage.block_cache;
-        if cache_cfg.shared && cache_cfg.capacity.is_none() {
-            cache_cfg.capacity = Some(ReadableSize {
-                0: self.rocksdb.defaultcf.block_cache_size.0
-                    + self.rocksdb.writecf.block_cache_size.0
-                    + self.rocksdb.lockcf.block_cache_size.0
-                    + self.raftdb.defaultcf.block_cache_size.0,
-            });
-        }
-    }
-
-    pub fn check_critical_cfg_with(&self, last_cfg: &Self) -> Result<(), String> {
-        if last_cfg.rocksdb.wal_dir != self.rocksdb.wal_dir {
-            return Err(format!(
-                "db wal_dir have been changed, former db wal_dir is '{}', \
-                 current db wal_dir is '{}', please guarantee all data wal logs \
-                 have been moved to destination directory.",
-                last_cfg.rocksdb.wal_dir, self.rocksdb.wal_dir
-            ));
-        }
-
-        if last_cfg.raftdb.wal_dir != self.raftdb.wal_dir {
-            return Err(format!(
-                "raftdb wal_dir have been changed, former raftdb wal_dir is '{}', \
-                 current raftdb wal_dir is '{}', please guarantee all raft wal logs \
-                 have been moved to destination directory.",
-                last_cfg.raftdb.wal_dir, self.rocksdb.wal_dir
-            ));
-        }
-
-        if last_cfg.storage.data_dir != self.storage.data_dir {
-            // In tikv 3.0 the default value of storage.data-dir changed
-            // from "" to "./"
-            let using_default_after_upgrade =
-                last_cfg.storage.data_dir.is_empty() && self.storage.data_dir == DEFAULT_DATA_DIR;
-
-            if !using_default_after_upgrade {
-                return Err(format!(
-                    "storage data dir have been changed, former data dir is {}, \
-                     current data dir is {}, please check if it is expected.",
-                    last_cfg.storage.data_dir, self.storage.data_dir
-                ));
-            }
-        }
-
-        if last_cfg.raft_store.raftdb_path != self.raft_store.raftdb_path {
-            return Err(format!(
-                "raft dir have been changed, former raft dir is '{}', \
-                 current raft dir is '{}', please check if it is expected.",
-                last_cfg.raft_store.raftdb_path, self.raft_store.raftdb_path
-            ));
-        }
-
-        Ok(())
-    }
-
-    pub fn from_file<P: AsRef<Path>>(path: P) -> Self {
-        (|| -> Result<Self, Box<dyn Error>> {
-            let s = fs::read_to_string(&path)?;
-            Ok(::toml::from_str(&s)?)
-        })()
-        .unwrap_or_else(|e| {
-            panic!(
-                "invalid auto generated configuration file {}, err {}",
-                path.as_ref().display(),
-                e
-            );
-        })
-    }
-
-    pub fn write_to_file<P: AsRef<Path>>(&self, path: P) -> Result<(), IoError> {
-        let content = ::toml::to_string(&self).unwrap();
-        let mut f = fs::File::create(&path)?;
-        f.write_all(content.as_bytes())?;
-        f.sync_all()?;
-
-        Ok(())
-    }
-
-    pub fn write_into_metrics(&self) {
-        self.raft_store.write_into_metrics();
-        self.rocksdb.write_into_metrics();
-    }
-}
-
-/// Prevents launching with an incompatible configuration
-///
-/// Loads the previously-loaded configuration from `last_tikv.toml`,
-/// compares key configuration items and fails if they are not
-/// identical.
-pub fn check_critical_config(config: &TiKvConfig) -> Result<(), String> {
-    // Check current critical configurations with last time, if there are some
-    // changes, user must guarantee relevant works have been done.
-    let store_path = Path::new(&config.storage.data_dir);
-    let last_cfg_path = store_path.join(LAST_CONFIG_FILE);
-
-    if last_cfg_path.exists() {
-        let last_cfg = TiKvConfig::from_file(&last_cfg_path);
-        config.check_critical_cfg_with(&last_cfg)?;
-    }
-
-    Ok(())
-}
-
-/// Persists critical config to `last_tikv.toml`
-pub fn persist_critical_config(config: &TiKvConfig) -> Result<(), String> {
-    let store_path = Path::new(&config.storage.data_dir);
-    let last_cfg_path = store_path.join(LAST_CONFIG_FILE);
-
-    // Create parent directory if missing.
-    if let Err(e) = fs::create_dir_all(&store_path) {
-        return Err(format!(
-            "create parent directory '{}' failed: {}",
-            store_path.to_str().unwrap(),
-            e
-        ));
-    }
-
-    // Persist current critical configurations to file.
-    if let Err(e) = config.write_to_file(&last_cfg_path) {
-        return Err(format!(
-            "persist critical config to '{}' failed: {}",
-            last_cfg_path.to_str().unwrap(),
-            e
-        ));
-    }
-
-    Ok(())
-}
-
-fn to_config_entry(change: ConfigChange) -> CfgResult<Vec<configpb::ConfigEntry>> {
-    // This helper function translate nested module config to a list
-    // of name/value pair and seperated module by '.' in the name field,
-    // by recursive call helper function with an prefix which represent
-    // th prefix of current module. And also compatible the field name
-    // in config struct with the name in toml file.
-    fn helper(prefix: String, change: ConfigChange) -> CfgResult<Vec<configpb::ConfigEntry>> {
-        let mut entries = Vec::with_capacity(change.len());
-        for (mut name, value) in change {
-            if name == "raft_store" {
-                name = "raftstore".to_owned();
-            } else {
-                name = name.replace("_", "-");
-            }
-            if !prefix.is_empty() {
-                let mut p = prefix.clone();
-                p.push_str(&format!(".{}", name));
-                name = p;
-            }
-            if let ConfigValue::Module(change) = value {
-                entries.append(&mut helper(name, change)?);
-            } else {
-                let mut e = configpb::ConfigEntry::default();
-                e.set_name(name);
-                e.set_value(from_change_value(value)?);
-                entries.push(e);
-            }
-        }
-        Ok(entries)
-    };
-    helper("".to_owned(), change)
-}
-
-fn from_change_value(v: ConfigValue) -> CfgResult<String> {
-    let s = match v {
-        ConfigValue::Duration(_) => {
-            let v: ReadableDuration = v.into();
-            toml::to_string(&v)?
-        }
-        ConfigValue::Size(_) => {
-            let v: ReadableSize = v.into();
-            toml::to_string(&v)?
-        }
-        ConfigValue::U64(ref v) => toml::to_string(v)?,
-        ConfigValue::F64(ref v) => toml::to_string(v)?,
-        ConfigValue::Usize(ref v) => toml::to_string(v)?,
-        ConfigValue::Bool(ref v) => toml::to_string(v)?,
-        ConfigValue::String(ref v) => toml::to_string(v)?,
-        _ => unreachable!(),
-    };
-    Ok(s)
-}
-
-/// Comparing two `Version` with the assumption of `global` and `local`
-/// should be monotonically increased, if `global` or `local` of _current config_
-/// less than _incoming config_ means there are update in _incoming config_
-pub fn cmp_version(current: &configpb::Version, incoming: &configpb::Version) -> Ordering {
-    match (
-        Ord::cmp(&current.local, &incoming.local),
-        Ord::cmp(&current.global, &incoming.global),
-    ) {
-        (Ordering::Equal, Ordering::Equal) => Ordering::Equal,
-        (Ordering::Less, _) | (_, Ordering::Less) => Ordering::Less,
-        _ => Ordering::Greater,
-    }
-}
-
-type CfgResult<T> = Result<T, Box<dyn Error>>;
-
-/// ConfigController use to register each module's config manager,
-/// and dispatch the change of config to corresponding managers or
-/// return the change if the incoming change is invalid.
-#[derive(Default)]
-pub struct ConfigController {
-    current: TiKvConfig,
-}
-
-impl ConfigController {
-    pub fn new(cfg: TiKvConfig) -> Self {
-        ConfigController { current: cfg }
-    }
-
-    fn update_or_rollback(&mut self, mut incoming: TiKvConfig) -> Option<ConfigChange> {
-        if incoming.validate().is_err() {
-            return Some(incoming.diff(&self.current));
-        }
-        let diff = self.current.diff(&incoming);
-        if !diff.is_empty() {
-            // TODO: dispatch change to each module
-            self.current.update(diff);
-        }
-        None
-    }
-
-    pub fn get_current(&self) -> &TiKvConfig {
-        &self.current
-    }
-}
-
-pub struct ConfigHandler {
-    id: String,
-    version: configpb::Version,
-    config_controller: ConfigController,
-}
-
-impl ConfigHandler {
-    pub fn start(
-        id: String,
-        controller: ConfigController,
-        version: configpb::Version,
-        _scheduler: FutureScheduler<PdTask>,
-    ) -> CfgResult<Self> {
-        // TODO: currently we can't handle RefreshConfig task, because
-        // PD have not implement such service yet.
-
-        // if let Err(e) = scheduler.schedule(PdTask::RefreshConfig) {
-        //     return Err(format!("failed to schedule refresh config task: {:?}", e).into());
-        // }
-        Ok(ConfigHandler {
-            id,
-            version,
-            config_controller: controller,
-        })
-    }
-
-    pub fn get_refresh_interval(&self) -> Duration {
-        Duration::from(self.config_controller.current.refresh_config_interval)
-    }
-
-    pub fn get_id(&self) -> String {
-        self.id.clone()
-    }
-
-    pub fn get_version(&self) -> &configpb::Version {
-        &self.version
-    }
-
-    pub fn get_config(&self) -> &TiKvConfig {
-        self.config_controller.get_current()
-    }
-}
-
-impl ConfigHandler {
-    /// Register the local config to pd
-    pub fn create(
-        id: String,
-        pd_client: Arc<impl PdClient>,
-        cfg: TiKvConfig,
-    ) -> CfgResult<(configpb::Version, TiKvConfig)> {
-        let cfg = toml::to_string(&cfg)?;
-        let version = configpb::Version::default();
-        let mut resp = pd_client.register_config(id, version, cfg)?;
-        match resp.get_status().get_code() {
-            StatusCode::Ok => {
-                let cfg: TiKvConfig = toml::from_str(resp.get_config())?;
-                Ok((resp.take_version(), cfg))
-            }
-            code => {
-                debug!("failed to register config"; "status" => ?code);
-                Err(format!("{:?}", resp).into())
-            }
-        }
-    }
-
-    /// Update the local config if remote config had been changed,
-    /// rollback the remote config if the change are invalid.
-    pub fn refresh_config(&mut self, pd_client: Arc<impl PdClient>) -> CfgResult<()> {
-        let mut resp = pd_client.get_config(self.get_id(), self.version.clone())?;
-        let version = resp.take_version();
-        match resp.get_status().get_code() {
-            StatusCode::NotChange => Ok(()),
-            StatusCode::WrongVersion if cmp_version(&self.version, &version) == Ordering::Less => {
-                let incoming: TiKvConfig = toml::from_str(resp.get_config())?;
-                if let Some(rollback_change) = self.config_controller.update_or_rollback(incoming) {
-                    debug!(
-                        "tried to update local config to an invalid config";
-                        "version" => ?resp.version
-                    );
-                    let entries = to_config_entry(rollback_change)?;
-                    self.update_config(version, entries, pd_client)?;
-                } else {
-                    info!("local config updated"; "version" => ?resp.version);
-                    self.version = version;
-                }
-                Ok(())
-            }
-            code => {
-                debug!(
-                    "failed to get remote config";
-                    "status" => ?code,
-                    "version" => ?version
-                );
-                Err(format!("{:?}", resp).into())
-            }
-        }
-    }
-
-    fn update_config(
-        &mut self,
-        version: configpb::Version,
-        entries: Vec<configpb::ConfigEntry>,
-        pd_client: Arc<impl PdClient>,
-    ) -> CfgResult<()> {
-        let mut resp = pd_client.update_config(self.get_id(), version, entries)?;
-        match resp.get_status().get_code() {
-            StatusCode::Ok => {
-                self.version = resp.take_version();
-                Ok(())
-            }
-            code => {
-                debug!("failed to update remote config"; "status" => ?code);
-                Err(format!("{:?}", resp).into())
-            }
-        }
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use tempfile::Builder;
-
-    use super::*;
-    use kvproto::configpb::Version;
-    use slog::Level;
-    use std::cmp::Ordering;
-    use toml;
-
-    #[test]
-    fn test_check_critical_cfg_with() {
-        let mut tikv_cfg = TiKvConfig::default();
-        let mut last_cfg = TiKvConfig::default();
-        assert!(tikv_cfg.check_critical_cfg_with(&last_cfg).is_ok());
-
-        tikv_cfg.rocksdb.wal_dir = "/data/wal_dir".to_owned();
-        assert!(tikv_cfg.check_critical_cfg_with(&last_cfg).is_err());
-
-        last_cfg.rocksdb.wal_dir = "/data/wal_dir".to_owned();
-        assert!(tikv_cfg.check_critical_cfg_with(&last_cfg).is_ok());
-
-        tikv_cfg.raftdb.wal_dir = "/raft/wal_dir".to_owned();
-        assert!(tikv_cfg.check_critical_cfg_with(&last_cfg).is_err());
-
-        last_cfg.raftdb.wal_dir = "/raft/wal_dir".to_owned();
-        assert!(tikv_cfg.check_critical_cfg_with(&last_cfg).is_ok());
-
-        tikv_cfg.storage.data_dir = "/data1".to_owned();
-        assert!(tikv_cfg.check_critical_cfg_with(&last_cfg).is_err());
-
-        last_cfg.storage.data_dir = "/data1".to_owned();
-        assert!(tikv_cfg.check_critical_cfg_with(&last_cfg).is_ok());
-
-        tikv_cfg.raft_store.raftdb_path = "/raft_path".to_owned();
-        assert!(tikv_cfg.check_critical_cfg_with(&last_cfg).is_err());
-
-        last_cfg.raft_store.raftdb_path = "/raft_path".to_owned();
-        assert!(tikv_cfg.check_critical_cfg_with(&last_cfg).is_ok());
-    }
-
-    #[test]
-    fn test_persist_cfg() {
-        let dir = Builder::new().prefix("test_persist_cfg").tempdir().unwrap();
-        let path_buf = dir.path().join(LAST_CONFIG_FILE);
-        let file = path_buf.as_path().to_str().unwrap();
-        let (s1, s2) = ("/xxx/wal_dir".to_owned(), "/yyy/wal_dir".to_owned());
-
-        let mut tikv_cfg = TiKvConfig::default();
-
-        tikv_cfg.rocksdb.wal_dir = s1.clone();
-        tikv_cfg.raftdb.wal_dir = s2.clone();
-        tikv_cfg.write_to_file(file).unwrap();
-        let cfg_from_file = TiKvConfig::from_file(file);
-        assert_eq!(cfg_from_file.rocksdb.wal_dir, s1.clone());
-        assert_eq!(cfg_from_file.raftdb.wal_dir, s2.clone());
-
-        // write critical config when exist.
-        tikv_cfg.rocksdb.wal_dir = s2.clone();
-        tikv_cfg.raftdb.wal_dir = s1.clone();
-        tikv_cfg.write_to_file(file).unwrap();
-        let cfg_from_file = TiKvConfig::from_file(file);
-        assert_eq!(cfg_from_file.rocksdb.wal_dir, s2.clone());
-        assert_eq!(cfg_from_file.raftdb.wal_dir, s1.clone());
-    }
-
-    #[test]
-    fn test_create_parent_dir_if_missing() {
-        let root_path = Builder::new()
-            .prefix("test_create_parent_dir_if_missing")
-            .tempdir()
-            .unwrap();
-        let path = root_path.path().join("not_exist_dir");
-
-        let mut tikv_cfg = TiKvConfig::default();
-        tikv_cfg.storage.data_dir = path.as_path().to_str().unwrap().to_owned();
-        assert!(persist_critical_config(&tikv_cfg).is_ok());
-    }
-
-    #[test]
-    fn test_keepalive_check() {
-        let mut tikv_cfg = TiKvConfig::default();
-        tikv_cfg.pd.endpoints = vec!["".to_owned()];
-        let dur = tikv_cfg.raft_store.raft_heartbeat_interval();
-        tikv_cfg.server.grpc_keepalive_time = ReadableDuration(dur);
-        assert!(tikv_cfg.validate().is_err());
-        tikv_cfg.server.grpc_keepalive_time = ReadableDuration(dur * 2);
-        tikv_cfg.validate().unwrap();
-    }
-
-    #[test]
-    fn test_block_size() {
-        let mut tikv_cfg = TiKvConfig::default();
-        tikv_cfg.pd.endpoints = vec!["".to_owned()];
-        tikv_cfg.rocksdb.defaultcf.block_size = ReadableSize::gb(10);
-        tikv_cfg.rocksdb.lockcf.block_size = ReadableSize::gb(10);
-        tikv_cfg.rocksdb.writecf.block_size = ReadableSize::gb(10);
-        tikv_cfg.rocksdb.raftcf.block_size = ReadableSize::gb(10);
-        tikv_cfg.raftdb.defaultcf.block_size = ReadableSize::gb(10);
-        assert!(tikv_cfg.validate().is_err());
-        tikv_cfg.rocksdb.defaultcf.block_size = ReadableSize::kb(10);
-        tikv_cfg.rocksdb.lockcf.block_size = ReadableSize::kb(10);
-        tikv_cfg.rocksdb.writecf.block_size = ReadableSize::kb(10);
-        tikv_cfg.rocksdb.raftcf.block_size = ReadableSize::kb(10);
-        tikv_cfg.raftdb.defaultcf.block_size = ReadableSize::kb(10);
-        tikv_cfg.validate().unwrap();
-    }
-
-    #[test]
-    fn test_parse_log_level() {
-        #[derive(Serialize, Deserialize, Debug)]
-        struct LevelHolder {
-            #[serde(with = "log_level_serde")]
-            v: Level,
-        }
-
-        let legal_cases = vec![
-            ("critical", Level::Critical),
-            ("error", Level::Error),
-            ("warning", Level::Warning),
-            ("debug", Level::Debug),
-            ("trace", Level::Trace),
-            ("info", Level::Info),
-        ];
-        for (serialized, deserialized) in legal_cases {
-            let holder = LevelHolder { v: deserialized };
-            let res_string = toml::to_string(&holder).unwrap();
-            let exp_string = format!("v = \"{}\"\n", serialized);
-            assert_eq!(res_string, exp_string);
-            let res_value: LevelHolder = toml::from_str(&exp_string).unwrap();
-            assert_eq!(res_value.v, deserialized);
-        }
-
-        let compatibility_cases = vec![("warn", Level::Warning)];
-        for (serialized, deserialized) in compatibility_cases {
-            let variant_string = format!("v = \"{}\"\n", serialized);
-            let res_value: LevelHolder = toml::from_str(&variant_string).unwrap();
-            assert_eq!(res_value.v, deserialized);
-        }
-
-        let illegal_cases = vec!["foobar", ""];
-        for case in illegal_cases {
-            let string = format!("v = \"{}\"\n", case);
-            toml::from_str::<LevelHolder>(&string).unwrap_err();
-        }
-    }
-
-    #[test]
-    fn test_config_change_to_config_entry() {
-        let old = TiKvConfig::default();
-        let mut incoming = TiKvConfig::default();
-        incoming.refresh_config_interval = ReadableDuration::hours(10);
-        let diff = to_config_entry(old.diff(&incoming)).unwrap();
-        assert_eq!(diff.len(), 1);
-        assert_eq!(diff[0].name, "refresh-config-interval");
-        assert_eq!(
-            diff[0].value,
-            toml::to_string(&incoming.refresh_config_interval).unwrap()
-        );
-    }
-
-    #[test]
-    fn test_cmp_version() {
-        fn new_version((g1, l1): (u64, u64), (g2, l2): (u64, u64)) -> (Version, Version) {
-            let mut v1 = Version::default();
-            v1.set_global(g1);
-            v1.set_local(l1);
-            let mut v2 = Version::default();
-            v2.set_global(g2);
-            v2.set_local(l2);
-            (v1, v2)
-        }
-
-        let (v1, v2) = new_version((10, 10), (10, 10));
-        assert_eq!(cmp_version(&v1, &v2), Ordering::Equal);
-
-        // either global or local of v1 less than global or local of v2
-        // Ordering::Less shuold be returned
-        let (small, big) = (10, 11);
-        let (v1, v2) = new_version((small, 10), (big, 10));
-        assert_eq!(cmp_version(&v1, &v2), Ordering::Less);
-        let (v1, v2) = new_version((small, 20), (big, 10));
-        assert_eq!(cmp_version(&v1, &v2), Ordering::Less);
-        let (v1, v2) = new_version((small, 10), (big, 20));
-        assert_eq!(cmp_version(&v1, &v2), Ordering::Less);
-
-        let (v1, v2) = new_version((10, small), (10, big));
-        assert_eq!(cmp_version(&v1, &v2), Ordering::Less);
-        let (v1, v2) = new_version((20, small), (10, big));
-        assert_eq!(cmp_version(&v1, &v2), Ordering::Less);
-        let (v1, v2) = new_version((10, small), (20, big));
-        assert_eq!(cmp_version(&v1, &v2), Ordering::Less);
     }
 }
